refactor plugin configuration classes to avoid CodeFactor complexity warning
# Conflicts:
#	tests/SharedPluginsAndCodeactivites/ChildPreCreatePlugin.cs
#	tests/SharedPluginsAndCodeactivites/ParentPostCreatePlugin.cs
#	tests/SharedPluginsAndCodeactivites/Plugin.cs
