+++ tests/SharedPluginsAndCodeactivites/ContactPostPlugin.cs
@@ -2,6 +2,7 @@
     using System;
     using Microsoft.Xrm.Sdk;
     using DG.XrmFramework.BusinessDomain.ServiceContext;
+    using static SharedPluginsAndCodeactivites.Utility.Enums;
 
     public class ContactPostPlugin : Plugin 
     {
