+++ tests/SharedPluginsAndCodeactivites/Plugin.cs
@@ -1,4 +1,5 @@
-﻿namespace DG.Some.Namespace {
+﻿namespace DG.Some.Namespace
+{
     using System;
     using System.Collections.Generic;
     using System.Collections.ObjectModel;
@@ -15,6 +16,7 @@
     using ExtendedStepConfig = System.Tuple<int, int, string, int, string, System.Guid?>;
     using ImageTuple = System.Tuple<string, string, int, string>;
     using System.Reflection;
+    using static SharedPluginsAndCodeactivites.Utility.Enums;
 
     /// <summary>
     /// Base class for all Plugins.
@@ -313,472 +315,7 @@
 
     }
 
-    #region PluginStepConfig made by Delegate A/S
-    interface IPluginStepConfig {
-        string _LogicalName { get; }
-        string _EventOperation { get; }
-        int _ExecutionStage { get; }
-
-        string _Name { get; }
-        int _Deployment { get; }
-        int _ExecutionMode { get; }
-        int _ExecutionOrder { get; }
-        string _FilteredAttributes { get; }
-        Guid? _UserContext { get; }
-        IEnumerable<ImageTuple> GetImages();
-    }
-
-    /// <summary>
-    /// Made by Delegate A/S
-    /// Class to encapsulate the various configurations that can be made 
-    /// to a plugin step.
-    /// </summary>
-    public class PluginStepConfig<T> : IPluginStepConfig where T : Entity
-    {
-        public string _LogicalName { get; private set; }
-        public string _EventOperation { get; private set; }
-        public int _ExecutionStage { get; private set; }
-
-        public string _Name { get; private set; }
-        public int _Deployment { get; private set; }
-        public int _ExecutionMode { get; private set; }
-        public int _ExecutionOrder { get; private set; }
-        public Guid? _UserContext { get; private set; }
-
-        public Collection<PluginStepImage> _Images = new Collection<PluginStepImage>();
-        public Collection<string> _FilteredAttributesCollection = new Collection<string>();
-
-        public string _FilteredAttributes
-        {
-            get
-            {
-                if (this._FilteredAttributesCollection.Count == 0) return null;
-                return string.Join(",", this._FilteredAttributesCollection).ToLower();
-            }
-        }
 
 
-        public PluginStepConfig(EventOperation eventOperation, ExecutionStage executionStage)
-        {
-            this._LogicalName = Activator.CreateInstance<T>().LogicalName;
-            this._EventOperation = eventOperation.ToString();
-            this._ExecutionStage = (int)executionStage;
-            this._Deployment = (int)Deployment.ServerOnly;
-            this._ExecutionMode = (int)ExecutionMode.Synchronous;
-            this._ExecutionOrder = 1;
-            this._UserContext = Guid.Empty;
-        }
 
-        private PluginStepConfig<T> AddFilteredAttribute(Expression<Func<T, object>> lambda)
-        {
-            this._FilteredAttributesCollection.Add(GetMemberName(lambda));
-            return this;
-        }
-
-        public PluginStepConfig<T> AddFilteredAttributes(params Expression<Func<T, object>>[] lambdas)
-        {
-            foreach (var lambda in lambdas) this.AddFilteredAttribute(lambda);
-            return this;
-        }
-
-        public PluginStepConfig<T> SetDeployment(Deployment deployment)
-        {
-            this._Deployment = (int)deployment;
-            return this;
-        }
-
-        public PluginStepConfig<T> SetExecutionMode(ExecutionMode executionMode)
-        {
-            this._ExecutionMode = (int)executionMode;
-            return this;
-        }
-
-        public PluginStepConfig<T> SetName(string name)
-        {
-            this._Name = name;
-            return this;
-        }
-
-        public PluginStepConfig<T> SetExecutionOrder(int executionOrder)
-        {
-            this._ExecutionOrder = executionOrder;
-            return this;
-        }
-
-        public PluginStepConfig<T> SetUserContext(Guid userContext)
-        {
-            this._UserContext = userContext;
-            return this;
-        }
-
-        public PluginStepConfig<T> AddImage(ImageType imageType)
-        {
-            return this.AddImage(imageType, null);
-        }
-
-        public PluginStepConfig<T> AddImage(ImageType imageType, params Expression<Func<T, object>>[] attributes)
-        {
-            return this.AddImage(imageType.ToString(), imageType.ToString(), imageType, attributes);
-        }
-
-        public PluginStepConfig<T> AddImage(string name, string entityAlias, ImageType imageType)
-        {
-            return this.AddImage(name, entityAlias, imageType, null);
-        }
-
-        public PluginStepConfig<T> AddImage(string name, string entityAlias, ImageType imageType, params Expression<Func<T, object>>[] attributes)
-        {
-            this._Images.Add(new PluginStepImage(name, entityAlias, imageType, attributes));
-            return this;
-        }
-
-        public IEnumerable<ImageTuple> GetImages()
-        {
-            foreach (var image in this._Images)
-            {
-                yield return new ImageTuple(image.Name, image.EntityAlias, image.ImageType, image.Attributes);
-            }
-        }
-
-        /// <summary>
-        /// Container for information about images attached to steps
-        /// </summary>
-        public class PluginStepImage
-        {
-            public string Name { get; private set; }
-            public string EntityAlias { get; private set; }
-            public int ImageType { get; private set; }
-            public string Attributes { get; private set; }
-
-            public PluginStepImage(string name, string entityAlias, ImageType imageType, Expression<Func<T, object>>[] attributes)
-            {
-                this.Name = name;
-                this.EntityAlias = entityAlias;
-                this.ImageType = (int)imageType;
-
-                if (attributes != null && attributes.Length > 0)
-                {
-                    this.Attributes = string.Join(",", attributes.Select(x => PluginStepConfig<T>.GetMemberName(x))).ToLower();
-                }
-                else
-                {
-                    this.Attributes = null;
-                }
-            }
-        }
-
-
-        private static string GetMemberName(Expression<Func<T, object>> lambda)
-        {
-            if (!(lambda.Body is MemberExpression body))
-            {
-                var ubody = (UnaryExpression)lambda.Body;
-                body = ubody.Operand as MemberExpression;
-            }
-
-            return body.Member.Name;
-        }
-    }
-
-    /// <summary>
-    /// Made by Delegate A/S
-    /// Class to encapsulate the various configurations that can be made 
-    /// to a plugin step.
-    /// </summary>
-    public class PluginStepConfig : IPluginStepConfig 
-    {
-        public string _LogicalName { get; private set; }
-        public string _EventOperation { get; private set; }
-        public int _ExecutionStage { get; private set; }
-
-        public string _Name { get; private set; }
-        public int _Deployment { get; private set; }
-        public int _ExecutionMode { get; private set; }
-        public int _ExecutionOrder { get; private set; }
-        public Guid? _UserContext { get; private set; }
-
-        public Collection<PluginStepImage> _Images = new Collection<PluginStepImage>();
-        public Collection<string> _FilteredAttributesCollection = new Collection<string>();
-
-        public string _FilteredAttributes
-        {
-            get
-            {
-                if (this._FilteredAttributesCollection.Count == 0) return null;
-                return string.Join(",", this._FilteredAttributesCollection).ToLower();
-            }
-        }
-
-        public PluginStepConfig(string logicalName,EventOperation eventOperation, ExecutionStage executionStage)
-        {
-            this._LogicalName = logicalName;
-            this._EventOperation = eventOperation.ToString();
-            this._ExecutionStage = (int)executionStage;
-            this._Deployment = (int)Deployment.ServerOnly;
-            this._ExecutionMode = (int)ExecutionMode.Synchronous;
-            this._ExecutionOrder = 1;
-            this._UserContext = Guid.Empty;
-        }
-
-        private PluginStepConfig AddFilteredAttribute(string name)
-        {
-            this._FilteredAttributesCollection.Add(name);
-            return this;
-        }
-
-        public PluginStepConfig AddFilteredAttributes(params string[] names)
-        {
-            foreach (var name in names) this.AddFilteredAttribute(name);
-            return this;
-        }
-
-        public PluginStepConfig SetDeployment(Deployment deployment)
-        {
-            this._Deployment = (int)deployment;
-            return this;
-        }
-
-        public PluginStepConfig SetExecutionMode(ExecutionMode executionMode)
-        {
-            this._ExecutionMode = (int)executionMode;
-            return this;
-        }
-
-        public PluginStepConfig SetName(string name)
-        {
-            this._Name = name;
-            return this;
-        }
-
-        public PluginStepConfig SetExecutionOrder(int executionOrder)
-        {
-            this._ExecutionOrder = executionOrder;
-            return this;
-        }
-
-        public PluginStepConfig SetUserContext(Guid userContext)
-        {
-            this._UserContext = userContext;
-            return this;
-        }
-
-        public PluginStepConfig AddImage(ImageType imageType)
-        {
-            return this.AddImage(imageType, null);
-        }
-
-        public PluginStepConfig AddImage(ImageType imageType, string[] attributes)
-        {
-            return this.AddImage(imageType.ToString(), imageType.ToString(), imageType, attributes);
-        }
-
-        public PluginStepConfig AddImage(string name, string entityAlias, ImageType imageType)
-        {
-            return this.AddImage(name, entityAlias, imageType, null);
-        }
-
-        public PluginStepConfig AddImage(string name, string entityAlias, ImageType imageType, string[] attributes)
-        {
-            this._Images.Add(new PluginStepImage(name, entityAlias, imageType, attributes));
-            return this;
-        }
-
-        public IEnumerable<ImageTuple> GetImages()
-        {
-            foreach (var image in this._Images)
-            {
-                yield return new ImageTuple(image.Name, image.EntityAlias, image.ImageType, image.Attributes);
-            }
-        }
-
-        /// <summary>
-        /// Container for information about images attached to steps
-        /// </summary>
-        public class PluginStepImage
-        {
-            public string Name { get; private set; }
-            public string EntityAlias { get; private set; }
-            public int ImageType { get; private set; }
-            public string Attributes { get; private set; }
-
-            public PluginStepImage(string name, string entityAlias, ImageType imageType, Expression<Func<object>>[] attributes)
-            {
-                this.Name = name;
-                this.EntityAlias = entityAlias;
-                this.ImageType = (int)imageType;
-
-                if (attributes != null && attributes.Length > 0)
-                {
-                    this.Attributes = string.Join(",", attributes.Select(x => PluginStepConfig.GetMemberName(x))).ToLower();
-                }
-                else
-                {
-                    this.Attributes = null;
-                }
-            }
-
-            public PluginStepImage(string name, string entityAlias, ImageType imageType, string[] attributes)
-            {
-                this.Name = name;
-                this.EntityAlias = entityAlias;
-                this.ImageType = (int)imageType;
-
-                if (attributes != null && attributes.Length > 0)
-                {
-                    this.Attributes = string.Join(",", attributes.Select(x => x.ToLower()));
-                }
-                else
-                {
-                    this.Attributes = null;
-                }
-            }
-        }
-
-        private static string GetMemberName(Expression<Func<object>> lambda)
-        {
-            if (!(lambda.Body is MemberExpression body))
-            {
-                var ubody = (UnaryExpression)lambda.Body;
-                body = ubody.Operand as MemberExpression;
-            }
-
-            return body.Member.Name;
-        }
-    }
-
-    class AnyEntity : Entity {
-        public AnyEntity() : base("") { }
-    }
-
-    /**
-     * Enums to help setup plugin steps
-     */
-
-    public enum ExecutionMode {
-        Synchronous = 0,
-        Asynchronous = 1,
-    }
-
-    public enum ExecutionStage {
-        PreValidation = 10,
-        PreOperation = 20,
-        PostOperation = 40,
-    }
-
-    public enum Deployment {
-        ServerOnly = 0,
-        MicrosoftDynamicsCRMClientforOutlookOnly = 1,
-        Both = 2,
-    }
-
-    // EventOperation based on CRM 2016
-    public enum EventOperation {
-        AddItem,
-        AddListMembers,
-        AddMember,
-        AddMembers,
-        AddPrincipalToQueue,
-        AddPrivileges,
-        AddProductToKit,
-        AddRecurrence,
-        AddToQueue,
-        AddUserToRecordTeam,
-        ApplyRecordCreationAndUpdateRule,
-        Assign,
-        AssignUserRoles,
-        Associate,
-        BackgroundSend,
-        Book,
-        CalculatePrice,
-        Cancel,
-        CheckIncoming,
-        CheckPromote,
-        Clone,
-        CloneProduct,
-        Close,
-        CopyDynamicListToStatic,
-        CopySystemForm,
-        Create,
-        CreateException,
-        CreateInstance,
-        CreateKnowledgeArticleTranslation,
-        CreateKnowledgeArticleVersion,
-        Delete,
-        DeleteOpenInstances,
-        DeliverIncoming,
-        DeliverPromote,
-        DetachFromQueue,
-        Disassociate,
-        Execute,
-        ExecuteById,
-        Export,
-        ExportAll,
-        ExportCompressed,
-        ExportCompressedAll,
-        GenerateSocialProfile,
-        GetDefaultPriceLevel,
-        GrantAccess,
-        Handle,
-        Import,
-        ImportAll,
-        ImportCompressedAll,
-        ImportCompressedWithProgress,
-        ImportWithProgress,
-        LockInvoicePricing,
-        LockSalesOrderPricing,
-        Lose,
-        Merge,
-        ModifyAccess,
-        PickFromQueue,
-        Publish,
-        PublishAll,
-        PublishTheme,
-        QualifyLead,
-        Recalculate,
-        ReleaseToQueue,
-        RemoveFromQueue,
-        RemoveItem,
-        RemoveMember,
-        RemoveMembers,
-        RemovePrivilege,
-        RemoveProductFromKit,
-        RemoveRelated,
-        RemoveUserFromRecordTeam,
-        RemoveUserRoles,
-        ReplacePrivileges,
-        Reschedule,
-        Retrieve,
-        RetrieveExchangeRate,
-        RetrieveFilteredForms,
-        RetrieveMultiple,
-        RetrievePersonalWall,
-        RetrievePrincipalAccess,
-        RetrieveRecordWall,
-        RetrieveSharedPrincipalsAndAccess,
-        RetrieveUnpublished,
-        RetrieveUnpublishedMultiple,
-        RetrieveUserQueues,
-        RevokeAccess,
-        Route,
-        RouteTo,
-        Send,
-        SendFromTemplate,
-        SetLocLabels,
-        SetRelated,
-        SetState,
-        SetStateDynamicEntity,
-        TriggerServiceEndpointCheck,
-        UnlockInvoicePricing,
-        UnlockSalesOrderPricing,
-        Update,
-        ValidateRecurrenceRule,
-        Win
-    }
-
-    public enum ImageType {
-        PreImage = 0,
-        PostImage = 1,
-        Both = 2,
-    }
-    #endregion
 }
