+++ tests/SharedPluginsAndCodeactivites/SharedPluginsAndCodeactivites.projitems
@@ -49,5 +49,10 @@
     <Compile Include="$(MSBuildThisFileDirectory)OpportunityWinLose.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PluginNonDaxif.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Plugin.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utility\AnyEntity.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utility\Enums.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utility\IPluginStepConfig.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utility\PluginStepConfig.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utility\TypedPluginStepConfig.cs" />
   </ItemGroup>
 </Project>
